Updated migration

+++ database/migrations/2022_12_07_091744_create_reviews_table.php
@@ -15,11 +15,11 @@ return new class extends Migration
     {
         Schema::create('reviews', function (Blueprint $table) {
             $table->id();
-			$table->integer('product_id')->unsigned()->index();
+			$table->bigInteger('product_id')->unsigned()->index();
 			$table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
 			$table->string('customer');
 			$table->text('review');
-			$table->integer('star');
+			$table->tinyInteger('star');
             $table->timestamps();
         });
     }
